handle sizeof(time_t) > sizeof(long) in format strings
Last-Update: 2024-03-11
Forwarded: no

64-bit time_t means that on some architectures, time_t is now larger than
a long, and making some references in format strings incorrect.  To avoid
truncation or other size mismatch issues, always cast to a long long and
read using %lld.

Fixes an assertion failure detected during build-time tests on armhf:
slapd: ../../../../../servers/slapd/overlays/dds.c:422: dds_op_add: Assertion `bv.bv_len < sizeof( ttlbuf )' failed.

Gbp-Pq: Name 64-bit-time-t-compat.patch

+++ contrib/slapd-modules/smbk5pwd/smbk5pwd.c
@@ -513,7 +513,7 @@ static int smbk5pwd_exop_passwd(
 		keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
 		keys[0].bv_len = snprintf(keys[0].bv_val,
 			LDAP_PVT_INTTYPE_CHARS(long),
-			"%ld", slap_get_time());
+			"%lld", (long long)slap_get_time());
 		BER_BVZERO( &keys[1] );
 		
 		ml->sml_desc = ad_sambaPwdLastSet;
@@ -535,7 +535,7 @@ static int smbk5pwd_exop_passwd(
 			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
 			keys[0].bv_len = snprintf(keys[0].bv_val,
 					LDAP_PVT_INTTYPE_CHARS(long),
-					"%ld", slap_get_time() + pi->smb_must_change);
+					"%lld", (long long)(slap_get_time() + pi->smb_must_change));
 			BER_BVZERO( &keys[1] );
 
 			ml->sml_desc = ad_sambaPwdMustChange;
@@ -558,7 +558,7 @@ static int smbk5pwd_exop_passwd(
 			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
 			keys[0].bv_len = snprintf(keys[0].bv_val,
 					LDAP_PVT_INTTYPE_CHARS(long),
-					"%ld", slap_get_time() + pi->smb_can_change);
+					"%lld", (long long)(slap_get_time() + pi->smb_can_change));
 			BER_BVZERO( &keys[1] );
 
 			ml->sml_desc = ad_sambaPwdCanChange;
